requirements.txt
Django
google-genai
python-decouple
gunicorn
psycopg2-binary
django-cors-headers